+++ NEWS.md
@@ -7,6 +7,7 @@ Version 0.11.2 - dev
 
 - NEW: `Insert.block()` returns associated `BlockLayout()` or `None` if block not exist or is an XREF
 - NEW: `Insert.has_scaling` returns `True` if any axis scaling is applied
+- NEW: `Insert.explode()` explode block reference entities into target layout
 - NEW: `UCS.scale()` creates a new scaled UCS
 - NEW: `Block` properties `is_anonymous`, `is_xref` and `is_xref_overlay`
 - CHANGE: renamed `Insert.ucs()` to `Insert.brcs()` which now returns a `BRCS()` object